requirements.txt
flask==2.3.3
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
python-telegram-bot==20.6
gunicorn==21.2.0 